clive/crm_chatbot/requirements.txt
flask>=2.3.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pymongo>=4.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
email-validator>=2.0.0
google-genai>=1.25.0 
uuid
datetime